오성 - 교수의 출석 저장 및 열람, 학생의 출석 열람

+++ src/main/java/com/project/hit/Attendance/AttendanceRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface AttendanceRepository extends JpaRepository<Attendance, Integer> {
+public interface AttendanceRepository extends JpaRepository<Attendance, Long> {
     List<Attendance> findAttendanceByDate(String date);
 }
